refactor(api)!: make the public sample rate u32 everywhere
BREAKING. `Options::sample_rate()` returned `f32` while `write_wav` and
`RealtimeOptions` took `u32`, so the documented quick-start had to write
`opt.sample_rate() as u32` at the seam between two halves of the same crate's
API. Sample rates are integers; the float was the odd one out.

Done now, deliberately, because ferrosintesis is not yet published: today this
costs nothing, and after the first release it is a major bump.

Changed to take/return `u32`:

    Options::with_sample_rate, Options::sample_rate
    loudness::{momentary_lufs, integrated_lufs, true_peak_dbtp, limit_true_peak}
    engine::normalize_loudness

All five loudness functions are re-exported from `offline`, so leaving them on
`f32` would have kept both types visible side by side in one module — which is
the actual complaint, not the single cast.

The private `Options::sr` field deliberately STAYS `f32`, and now carries a
comment saying so. `sr` feeds ~130 voice/filter/envelope constructors and is
divided into coefficients on the render path; making the field `u32` and
converting at each use would move the conversions onto that path and reorder
float operations. Byte-identical renders depend on this staying as it is.

Each converted function shadows its parameter (`let fs = fs as f32;`) as the
first statement, leaving every interior expression untouched.
`normalize_loudness` needs no shadow at all — its `sr` is never used in
arithmetic, only forwarded to two functions that now take `u32`.

VERIFICATION — a pure type change, so any render difference is a bug:

  tools/render-diff/render_diff.py, 124 MIDIs, baseline built from the current
  origin/main (rebuilt, not reused — the trunk moved twice today):
    EXPECTED changed :   0
    EXPECTED same    : 124
    CONTAMINATION    :   0
    NOT REACHED      :   0

  cargo test -p ferrosintesis --lib --release: 674 passed, 0 failed
  cargo clippy --workspace --exclude amp-lab --all-targets: clean
  cargo fmt --all --check: clean

Docs updated with the API: the README quick-start loses its `as u32`, the
options table records the `u32` type, and the crate-level doc comment stops
referring to the private `opt.sr`.

+++ crates/ferrosintesis-cli/examples/calmeter.rs
@@ -200,7 +200,7 @@ fn measure(w: &Wav, start: i64, window_frames: i64, preroll_frames: i64) -> (Vec
         .iter()
         .fold(0.0f32, |m, v| m.max(v.abs()));
 
-    let m = momentary_lufs(slice, w.sr as f32);
+    let m = momentary_lufs(slice, w.sr);
     // Block j covers [j*hop, j*hop + 400ms) inside the slice. Keep only blocks that
     // start at or after the true window origin — everything earlier is filter warm-up.
     let hop = 0.100 * w.sr as f64;
